image-source: Fix slideshow source memory usage
Now the slideshow source will only load 5 images at a time, with
each image is loaded and deleted when needed. I changed the
random to a shuffle because choosing a random file would
cause the image to be loaded on the spot. With these changes
the images are loaded two images ahead of time.